fix the compile options to apply to the miopen plugin correctly

+++ plugins/miopen_legacy_plugin/miopen_utils.hpp
@@ -33,7 +33,7 @@ namespace miopen_legacy_plugin
 namespace miopen_utils
 {
 
-hipdnnPluginDeviceBuffer_t find_device_buffer(int uid,
+hipdnnPluginDeviceBuffer_t find_device_buffer(int64_t uid,
                                               const hipdnnPluginDeviceBuffer_t* device_buffers,
                                               uint32_t num_device_buffers);
